# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: New Face Global Network API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: userId
          in: query
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderListResponse"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create order from cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderBody"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderStatusBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  # COMMISSIONS
...
  /messages:
    get:
      operationId: listMessages
      tags: [messages]
      summary: List messages
      parameters:
        - name: folder
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageListResponse"
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageBody"
      responses:
        "201":
          description: Sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
  /messages/{id}/read:
    post:
      operationId: markMessageRead
      tags: [messages]
      summary: Mark message as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Marked read
  /messages/broadcast:
    post:
      operationId: broadcastMessage
      tags: [messages]
      summary: Broadcast message to all members (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BroadcastMessageBody"
      responses:
        "201":
          description: Broadcast sent

  # SETTINGS
...
components:
  schemas:
...
    Order:
      type: object
      properties:
        id:
          type: integer
        orderNumber:
          type: string
        userId:
          type: integer
        userName:
          type: string
        status:
          type: string
        paymentMethod:
          type: string
        paymentStatus:
          type: string
        subtotal:
          type: number
        tax:
          type: number
        shipping:
          type: number
        total:
          type: number
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        shippingAddress:
          type: ["string", "null"]
        promoCode:
          type: ["string", "null"]
        discount:
          type: number
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - orderNumber
        - userId
        - userName
        - status
        - paymentMethod
        - paymentStatus
        - subtotal
        - tax
        - shipping
        - total
        - items
        - discount
        - createdAt
    OrderItem:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        productImage:
          type: ["string", "null"]
        price:
          type: number
        quantity:
          type: integer
        total:
          type: number
      required:
        - id
        - productId
        - productName
        - price
        - quantity
        - total
    OrderListResponse:
      type: object
      properties:
        orders:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required:
        - orders
        - total
        - page
        - totalPages
    CreateOrderBody:
      type: object
      properties:
        paymentMethod:
          type: string
        shippingAddress:
          type: string
        promoCode:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
      required:
        - paymentMethod
        - shippingAddress
    UpdateOrderStatusBody:
      type: object
      properties:
        status:
          type: string
        paymentStatus:
          type: ["string", "null"]
      required:
        - status
...
    Message:
      type: object
      properties:
        id:
          type: integer
        fromUserId:
          type: ["integer", "null"]
        fromUserName:
          type: string
        toUserId:
          type: ["integer", "null"]
        subject:
          type: string
        body:
          type: string
        isRead:
          type: boolean
        isBroadcast:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - fromUserName
        - subject
        - body
        - isRead
        - isBroadcast
        - createdAt
    MessageListResponse:
      type: object
      properties:
        messages:
          type: array
          items:
            $ref: "#/components/schemas/Message"
        total:
          type: integer
        unreadCount:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required:
        - messages
        - total
        - unreadCount
        - page
        - totalPages
    SendMessageBody:
      type: object
      properties:
        toUserId:
          type: integer
        subject:
          type: string
        body:
          type: string
      required:
        - toUserId
        - subject
        - body
    BroadcastMessageBody:
      type: object
      properties:
        subject:
          type: string
        body:
          type: string
        targetRole:
          type: ["string", "null"]
      required:
        - subject
        - body
...